fix opting out from displaying error in firefox private mode
It seems that rejecting dexie promises do weird things when
wrapped in a Promise.all. This caused Firefox in private
mode to display an error message on opting out when data was still
being deleted. Instead of running the purge operation in parallel
we can resort to serially pruning the stores.