Replace 'go-gomail/gomail' with 'wneessen/go-mail'
This commit replaces the deprecated 'go-gomail/gomail' package with 'wneessen/go-mail'. Additionally it adds `Authtype` as config option. Be default this is set to `LOGIN`, but can be either of `LOGIN`, `PLAIN` or `CRAM-MD5`. Since 'wneessen/go-mail' supports local sendmail delivery out of the box, some of the pervious sendmail code has been thrown away.